Update translation.json
There's one issue I'm not sure about concerning the translation: Will the syntax for adding a @context, +project, due:... be English when typing it in or will it be translated as well? Same goes for adding a date. The English way would be yyyy-mm-dd whereas the German format would be dd-mm-yyyy.

"search": "(A) Suche mithilfe der todo.txt Syntax @kontext (=german) +projekt (=german) due: (english)",

If I'm not mistaken, the Syntax needs to be used in English, even though the language of the System, thus, of the software is German.

I can standardize it throughout the whole document, however, I'd need to know what applies. :)